Add explicit cast to int for open_index() call
The getShardIndex() method might return size_t which would match the
deprecated open_index(name, size_t, bool, bool) signature instead of
the newer open_index(name, int) signature. Adding explicit cast to
ensure we call the correct overload.

+++ src/mako/benchmarks/ycsb.cc
@@ -384,8 +384,9 @@ public:
     : bench_runner(db)
   {
     // Use the new open_index(name, shard_index) API instead of deprecated open_index(name, value_size_hint)
+    // Cast to int to ensure we call the right overload
     auto& benchConfig = BenchmarkConfig::getInstance();
-    open_tables["USERTABLE"] = db->open_index("USERTABLE", benchConfig.getShardIndex());
+    open_tables["USERTABLE"] = db->open_index("USERTABLE", static_cast<int>(benchConfig.getShardIndex()));
   }
 
 protected:
